perf: Semi-NCA dominator algorithm + inboundOffsets int[] + postOrder GC fix
Three independent memory reductions targeting the DOM phase peak (~-8 GB combined
on 513M-object heaps):

1. Semi-NCA replaces Lengauer-Tarjan: eliminates bucket[] and next[] arrays
   (2 × int[reachable]); idomD and sdom retained for semi-dominator computation,
   NCA forward-pass replaces LT bucket-propagation. Saves ~4 GB at DOM peak.

2. inboundOffsets long[]→int[]: VByte stream offsets stored as unsigned int[]
   (Integer.toUnsignedLong() at read sites). Stream max ~2.5 GB < 4.29 GB uint32
   max on any supported heap. Saves ~2 GB at DOM peak.

3. Drop postOrder donation from RpoDfs phaseArrays: postOrder[N] was donated but
   immediately evicted by dfsOrder donation inside DominatorTree, sitting idle for
   the entire LT main loop. Removing the donation lets GC collect it. Saves ~2 GB.

All 245 tests pass. Validated on scala-doku (1,509 GC roots, ~20 MB Thread retained)
and VSCode 11M-object dump.

+++ src/main/java/me/bechberger/hprof/views/CsrBuilder.java
@@ -139,7 +139,7 @@ final class CsrBuilder {
      *
      * During addEdge, if excluded: store src | Integer.MIN_VALUE in targets[pos].
      * During encodeVByte, extract the flag from the sign bit before encoding.
-     * Sets graph.inboundOffsets (long[]) and graph.inboundStream (byte[][]) directly.
+     * Sets graph.inboundOffsets (int[]) and graph.inboundStream (byte[][]) directly.
      */
     void encodeVByteWithEmbeddedFlags() {
         int[] offsets = offsetsCursor;
@@ -150,7 +150,7 @@ final class CsrBuilder {
 
         // Estimate stream size; small heaps use single byte[], large use chunked byte[][].
         long estimatedBytes = Math.max((long) totalEdges + totalEdges / 4L, 16L);
-        long[] byteOffsets = new long[n + 1];
+        int[] byteOffsets = new int[n + 1];
         long streamPos = 0;
 
         if (estimatedBytes < VByte.CHUNK_SIZE) {
@@ -165,7 +165,7 @@ final class CsrBuilder {
                     if (rowLen > sortScratch.length) sortScratch = new long[rowLen];
                     sortWithFlags(targets, rowStart, rowEnd, sortScratch);
                 }
-                byteOffsets[v] = streamPos;
+                byteOffsets[v] = (int) streamPos;
                 int prev = 0;
                 for (int i = rowStart; i < rowEnd; i++) {
                     int rawSrc = targets[i];
@@ -181,7 +181,7 @@ final class CsrBuilder {
                     logicalEdgeIdx++;
                 }
             }
-            byteOffsets[n] = streamPos;
+            byteOffsets[n] = (int) streamPos;
             targets = null; offsetsCursor = null;
             if ((int) streamPos < singleBuf.length) singleBuf = Arrays.copyOf(singleBuf, (int) streamPos);
             graph.inboundStream  = new byte[][] { singleBuf };
@@ -206,7 +206,7 @@ final class CsrBuilder {
                 sortWithFlags(targets, rowStart, rowEnd, sortScratch);
             }
 
-            byteOffsets[v] = streamPos;
+            byteOffsets[v] = (int) streamPos;
             int prev = 0;
 
             for (int i = rowStart; i < rowEnd; i++) {
@@ -231,7 +231,7 @@ final class CsrBuilder {
             }
         }
 
-        byteOffsets[n] = streamPos;
+        byteOffsets[n] = (int) streamPos;
         targets = null;
         offsetsCursor = null;
